I figured out how to make Expressions with parameters work. The original plan was to copy boundary conditions including their parameters (which would require work on the Dolfin side), but that wouldn't work if the Expression showed up in the weak form itself.
Instead, we overload Expression.__init__ and Expression.__setattr__
to /record/ the values of all parameters as they are set. When an
equation is annotated, we /freeze/ the dictionary of parameters;
then the callbacks for the operators /update/ the parameters of
all of the Expressions with the values recorded during the annotation
when they are called.

That way, the operators always assemble things with the parameters
of the expressions set to the values they should be.

Activate the time-dependent boundary condition in the Navier-Stokes
test (i.e., update the t parameter). The forward replay and adjoint
tests still converge.