Test the functional gradient (fg) with respect to a scalar (Constant) parameter. It converges at second order, which is nice.
The clean way to do this is to cast the parameter as a Parameter
object, and then have that know how to use the adjoint to compute
the gradient with respect to the parameter. There will be some
refactoring next, but I thought I would commit this test as I have
it working.