At various points, dolfin-adjoint makes new boundary conditions. In order to stop the list of BCs growing and growing indefinitely, it makes the list of BCs unique at various points, which keeps the size of the list bounded. Previously, the way it made the list unique was the following:
  bcs = list(set(bcs))

However -- it turns out that this is not safe in parallel!! set() is unordered
and the order you happen to get isn't the same across processors. So if you have
a list

bcs = [DirichletBC(), PeriodicBC()]

and then apply list(set(bcs)), on one process the DirichletBC is first, and
on the other process the PeriodicBC is first, causing a deadlock.

Subtle, no?

The fix is to implement a smarter uniq() which removes duplicates from
a list in such a way that it preserves order.