a) Implement the last term in the separable second-order adjoint setup: the <foa, d2F/dm2 mdot> term. This is taken care of by the parameter.equation_partial_second_derivative method.
b) Change curl-curl to depend nonlinearly on the ScalarParameter, to have a test where a) is necessary.

c) Unfortunately, the caching Hessian wasn't always giving the correct answer. I think this is because
(adjoint matrix after boundary condition application) isn't the transpose of (tlm matrix after boundary
condition application): after BC application both will have *rows* homogenised, but transposing that
homogenises a *column*. So I changed it to factor both the TLM and ADM systems, which doubles the
storage cost and the upfront cost of the caching Hessian, which sucks. I'm sure there's a much smarter
way of handling this. But until I figure out what that should be, this will do.