Implement a caching strategy for the ReducedFunctional. It can be activated with cache="cache_file")
The idea is to "memoise" the ReducedFunctional. That is, before a reduced functional (or gradient/Hessian) invokes running the forward/adjoint systems,
it checks if the result is already stored in the memoisation dictionary. If not, it runs the forward/adjoint system and adds the result to the memoisation dictionary.

The current implementation only caches the functional evaluation.