A new option for minimize/maximize "in_euclidian_space" which internally transforms the solutions such that the inner product is the standard dot product.
This fixes the issue that until now the scipy algorithms do not accept to pass in the L2 inner product (which would be the right one for function optimisation parameters).

The current solutions relies on a Cholesky factorisation that is performed densly with scipy. Hence the current code does not scale to big problems and this is why the default setting
has not been changed.